Modificacion stock por errores

+++ ms-stock/src/main/java/cl/pchardware/stock/dto/BodegaResponse.java
@@ -1,11 +1,9 @@
 package cl.pchardware.stock.dto;
-
 import lombok.Data;
 
 @Data
 public class BodegaResponse {
-
-    private Integer idBodega;
+    private Long idBodega;
     private String codigo;
     private String nombre;
     private String tipo;
